requirements.txt
# Core dependencies - Production pinned versions
pandas==2.1.4
loguru==0.7.2
pydantic==2.5.2
streamlit==1.29.0
plotly==5.17.0
python-docx==1.1.0

# Log parsing and forensics
python-evtx==0.7.4
dateparser==1.2.0

# AI/ML and LLM integration
ibm-watson-machine-learning==1.0.335
groq==0.4.2
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20

# Vector database and retrieval
chromadb==0.4.22
sentence-transformers==2.2.2

# Security tools
vt-py==0.17.5
yara-python==4.3.1
scapy==2.5.0

# Network visualization
pyvis==0.3.2
networkx==3.2.1

# Additional utilities
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
boto3==1.34.0

# Production security and monitoring
cryptography==41.0.8
python-multipart==0.0.6